Logistique_Pro/requirements.txt
streamlit>=1.44.0
pandas>=2.2.2
bcrypt>=4.1.2
python-dotenv>=1.0.1
qrcode[pil]>=7.4.2
Pillow>=10.3.0
plotly>=5.22.0
openpyxl>=3.1.2
streamlit-autorefresh>=1.0.1
